requirements.txt
fastapi
pydantic
uvicorn
psycopg2-binary
cinemagoer
